+++ src/Generic/GenericCrudModel.php
@@ -7,7 +7,6 @@ use Rhubarb\Leaf\Crud\Leaves\CrudModel;
 class GenericCrudModel extends CrudModel
 {
     public $errors = [];
-
-    /** @var boolean */
-    public $displayActionButtons;
+    public $showButtonBar = true;
+    public $showHeader = true;
 }