+++ src/Generic/GenericItemView.php
@@ -37,19 +37,38 @@ abstract class GenericItemView extends CrudView
 
         print <<<HTML
         <main class="c-main">
-            {$this->getHeaderBarHTML()}
-            <div class="u-pad o-wrap">
+HTML;
+        if($this->model->showHeader) {
+            $this->getHeaderBarHTML();
+        }
+        print <<<HTML
+            <div
+} class="u-pad o-wrap">
                 <h1 class="u-epsilon c-heading">{$htmlPageSettings->pageTitle}</h1>
                 
 HTML;
 
         $this->printInputs();
 
-        $this->printActionButtons();
+        if($this->model->showButtonBar)
+        {
+            $this->printButtonBar();
+        }
 
         print <<<HTML
-        </div>
-    </main>
+            </div>
+        </main>
+HTML;
+    }
+
+    protected function printButtonBar()
+    {
+        $delete = $this->model->restModel->isNewRecord() ? '' : $this->leaves['Delete'];
+        print <<<HTML
+        <div class="u-pad-top-bottom o-flex">
+                    <div class="o-flex__item">{$this->leaves['Save']}{$this->leaves['Cancel']}</div>
+                    <div>{$delete}</div>
+                </div>
 HTML;
     }